requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
env==0.1.0
Markdown==3.7
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.2
